Complete renderForm function
The renderForm function is now complete, it renders a label, input,
select, and button as part of a form for submitting a location to get
the weather of. It is not connected at all, this is just for creating
the html and I added a bit of styling, mostly grid stuff for alignment.