Create bindEvents and connect ui to api call
I made a bindEvents file to handle my event listeners. Now you can
choose a location and unit of measurement and receive weather data from
the api call. Just in the console, but it works. I added error messages
for when the user enters a bad location or there is a generic error.